requirements.txt
requests==2.20.0
pytest==6.2.2
PyYAML==5.4.1
